refactor: move brush related const enum interface to seperated file
