+++ packages/mondrian/src/plugin/eraser-plugin.ts
@@ -5,7 +5,7 @@ import {
   IMondrianStateData,
   MondrianDataType,
 } from "../data-manager";
-import { BrushName } from "./brush-plugin";
+import { BrushName } from "./brush-common";
 import { PencilBrushPlugin } from "./pencil-plugin";
 import { PluginType } from "./plugin";
