+++ packages/mondrian/src/plugin/sphere-plugin.ts
@@ -6,7 +6,7 @@ import {
   IMondrianStateData,
   MondrianDataType,
 } from "../data-manager";
-import { BrushName } from "./brush-plugin";
+import { BrushName } from "./brush-common";
 import { PluginType } from "./plugin";
 import { ShapePlugin } from "./shape-plugin";
