+++ packages/mondrian/src/plugin/stroke-plugin.ts
@@ -4,7 +4,7 @@ import {
   MondrianDataType,
   IMondrianInteractData,
 } from "../data-manager";
-import { BrushName } from "./brush-plugin";
+import { BrushName } from "./brush-common";
 import { PluginType } from "./plugin";
 import { ShapePlugin } from "./shape-plugin";
